# restfull/FEMS/swagger.yaml
openapi: 3.0.3
info:
  title: TZW LTD Fire Extinguisher Management API
  version: 1.0.0
  description: |
    Consolidated API documentation for the FEMS backend services behind the API gateway.

    Covered services:
    - Authentication and user administration
    - Facility and customer management
    - Fire extinguisher registration and tracking
    - Inspection scheduling and recording
    - Maintenance logging
    - Notifications and escalations
    - Reporting and audit access
servers:
  - url: http://localhost:5000/api
    description: Local API gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit10Param:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Limit20Param:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
        default: ''
...
  schemas:
...
    CustomerCreateRequest:
      type: object
      required: [fullName, phone]
      properties:
        fullName:
          type: string
        nationalId:
          type: string
          nullable: true
        phone:
          type: string
          example: '+250788000000'
        email:
          type: string
          format: email
          nullable: true
        address:
          type: string
          nullable: true
        organizationName:
          type: string
          nullable: true
...
    Inspection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        extinguisherId:
          type: string
          format: uuid
        extinguisherCode:
          type: string
        inspectorId:
          type: string
          format: uuid
        inspectorName:
          type: string
        inspectionDate:
          type: string
          format: date
        inspectionTime:
          type: string
          nullable: true
          example: '10:30'
        findings:
          type: string
          nullable: true
        status:
          type: string
          enum: [Scheduled, Completed, Requires Service, Failed]
        nextInspectionDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
    InspectionCreateRequest:
      type: object
      required: [extinguisherId, inspectorId, inspectionDate, status]
      properties:
        extinguisherId:
          type: string
          format: uuid
        inspectorId:
          type: string
          format: uuid
        inspectionDate:
          type: string
          format: date
        inspectionTime:
          type: string
          nullable: true
          example: '10:30'
        findings:
          type: string
          nullable: true
        status:
          type: string
          enum: [Scheduled, Completed, Requires Service, Failed]
        nextInspectionDate:
          type: string
          format: date
          nullable: true
...
    MaintenanceCreateRequest:
      type: object
      required:
        [extinguisherId, serviceDate, serviceCompany, technicianName, actionTaken, issuesIdentified]
      properties:
        extinguisherId:
          type: string
          format: uuid
        serviceDate:
          type: string
          format: date
        serviceCompany:
          type: string
        technicianName:
          type: string
        actionTaken:
          type: string
        issuesIdentified:
          type: string
        recommendations:
          type: string
          nullable: true
        nextServiceDate:
          type: string
          format: date
          nullable: true
        cost:
          type: number
          default: 0
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
          default: completed
...
    ReportMeta:
      type: object
      properties:
        institution:
          type: string
        system:
          type: string
        title:
          type: string
        generatedBy:
          type: string
        generatedByEmail:
          type: string
        generatedByRole:
          type: string
        generatedAt:
          type: string
          format: date-time
        generatedAtDisplay:
          type: string
        totalRecords:
          type: integer
        filters:
          type: object
          additionalProperties: true
    ReportResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        report:
          $ref: '#/components/schemas/ReportMeta'
        columns:
          type: array
          items:
            type: string
        data:
          type: array
          items:
            type: object
            additionalProperties: true
security:
  - bearerAuth: []
paths:
...
  /customers:
    post:
      tags: [Customers]
      summary: Create a customer or facility record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateRequest'
      responses:
        '201':
          description: Customer created
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit10Param'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: Customer list
...
  /inspections:
    post:
      tags: [Inspections]
      summary: Schedule or record an inspection
      description: |
        Users can schedule only.
        Admins and inspectors can also record completed or failed outcomes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionCreateRequest'
      responses:
        '201':
          description: Inspection scheduled or recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  data:
                    $ref: '#/components/schemas/Inspection'
        '403':
          description: User role cannot submit the requested outcome
    get:
      tags: [Inspections]
      summary: List inspections
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit10Param'
        - name: extinguisherId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [Scheduled, Completed, Requires Service, Failed]
      responses:
        '200':
          description: Inspection list
...
  /maintenance:
    post:
      tags: [Maintenance]
      summary: Create a maintenance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceCreateRequest'
      responses:
        '201':
          description: Maintenance record created
    get:
      tags: [Maintenance]
      summary: List maintenance records
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit10Param'
        - name: extinguisherId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
      responses:
        '200':
          description: Maintenance list
...
  /reports/expired:
    get:
      tags: [Reports]
      summary: Expired extinguishers report
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports/expiring-soon:
    get:
      tags: [Reports]
      summary: Upcoming expirations report
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 90
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
  /reports/customers:
    get:
      tags: [Reports]
      summary: Facility report
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
  /reports/inspections:
    get:
      tags: [Reports]
      summary: Inspection report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
...
  /reports/maintenance:
    get:
      tags: [Reports]
      summary: Maintenance report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
...
  /reports/compliance:
    get:
      tags: [Reports]
      summary: Compliance report
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
...
  /reports/audit:
    get:
      tags: [Reports]
      summary: Audit log report
      description: Admin only
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit20Param'
        - name: action
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Audit trail list
...